refactor: cardDeck 생성 책임 위치 변경
