+++ src/test/java/domain/card/CardDeckTest.java
@@ -1,8 +1,7 @@
-package domain;
+package domain.card;
 
 import static domain.participant.Dealer.DECK_SIZE;
 
-import domain.card.CardDeck;
 import exception.NoMoreCardException;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
@@ -24,10 +23,10 @@ class CardDeckTest {
     @DisplayName("카드가 없는데 카드를 뽑을 경우 예외가 발생한다.")
     void pop() {
         //given
-        CardDeck cardDeck = CardDeck.generate(DECK_SIZE);
+        CardDeck cardDeck = CardDeck.generate(1);
 
         //when
-        int cardSize = 52 * 6;
+        int cardSize = 52;
         while (cardSize > 0) {
             cardDeck.pop();
             cardSize--;
